Several updates, see desc
These updates are to increase the max possible polling rate. Claude rewrote code to have faster SD logging (keeps file open and flushes every so often to not lose tons of data on power off). I rewrote battery check to only happen every 10 seconds to not waste time. Serial printing is now at 460800 instead of 115200 to reduce the time to print to Serial monitor (which is always happening) by 4x